Split tests, add order by
